+++ cpp/CMakeLists.txt
@@ -11,5 +11,5 @@ SET(WITH_STATIC ON CACHE BOOL "WITH_STATIC Indigo option")
 
 add_subdirectory(app)
 add_subdirectory(fingerprints)
-add_subdirectory(lib)
+add_subdirectory(qtrlib)
 add_subdirectory(third_party/Indigo)
