+++ cpp/app/CMakeLists.txt
@@ -10,7 +10,7 @@ target_include_directories(app
         )
 
 target_link_directories(app PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/../build/lib)
-target_link_libraries(app indigo-cpp qtr-libs)
+target_link_libraries(app indigo-cpp qtrlib)
 
 include(${CMAKE_CURRENT_SOURCE_DIR}/../cmake/SetOutputDirectory.cmake)
 set_output_directory(app)
