+++ cpp/qtrlib/profiler/SearchEngineProfiler.cpp
