+++ cpp/qtrlib/search_engine/BingoSearchEngine.cpp
