+++ cpp/qtrlib/search_engine/MockSearchEngine.cpp
