+++ cpp/qtrlib/search_engine/SearchEngineFactory.h
