+++ cpp/qtrlib/search_engine/SearchEngineInterface.h
