+++ cpp/qtrlib/utils/Utils.h
